requirements.txt
fastapi
uvicorn
python-multipart
pdfplumber
pandas
scipy
anthropic
python-dotenv
